moa_gateway: extract progress streaming into its own module
PR review (P3, James): moa_gateway.rs grew to 3,020 lines as the
streaming progress UX accreted around the gateway entry, well over
the repo's 2,000-line file limit. Split the progress-streaming
surface into a dedicated progress submodule so the gateway file can
stay focused on routing, scoring, and worker dispatch.

The progress module owns:
- The streaming entry (run_moa_turn_with_progress) and the ticker
  drip loop (drip_progress_phase, drip_progress_until_moa_completes,
  drip_progress_against_future).
- All progress-specific wire writers (write_progress_response_created,
  send_progress_headers, write_progress_event, write_progress_body,
  write_failure_as_sse_tail).
- The Responses-API continuation type (ProgressContinuation,
  ClientGone) and the small helpers tied to the progress path
  (item_id_from_completion_id, overwrite_response_id,
  progress_line, short_hex_nanos).
- All progress-related tests.

The gateway module keeps the body writers (send_moa_as_*_sse_inner)
and shared helpers (is_moa_failure_body, write_sse_response_headers,
chunking, MoA decision logic). Helpers needed by progress are
re-scoped to pub(in crate::network::openai::moa_gateway) so the
visibility boundary matches the module tree.

mod.rs: 1,953 lines. progress.rs: 1,108 lines. Both now under
the 2,000-line limit. No behaviour change \u2014 all 53 existing tests
pass without modification (just relocated into the progress test
module where they live next to the code they exercise).
